backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[argon2]
argon2-cffi
python-dotenv
httpx
pydantic
pydantic[email]
python-multipart
alembic
